+++ packages/api/endpoints/rules.js
@@ -10,13 +10,18 @@ const {
   RulePgModel,
   translateApiRuleToPostgresRule,
   translatePostgresRuleToApiRule,
+  translateApiRuleToPostgresRuleRaw,
   isCollisionError,
 } = require('@cumulus/db');
 const { Search } = require('@cumulus/es-client/search');
 const { indexRule, deleteRule } = require('@cumulus/es-client/indexer');
 
 const { isBadRequestError } = require('../lib/errors');
-const { deleteRuleResources } = require('../lib/rulesHelpers');
+const {
+  deleteRuleResources,
+  invokeRerun,
+  updateRuleTrigger,
+} = require('../lib/rulesHelpers');
 const models = require('../models');
 
 const log = new Logger({ sender: '@cumulus/api/rules' });
@@ -128,7 +133,6 @@ async function post(req, res) {
  */
 async function put(req, res) {
   const {
-    ruleModel = new models.Rule(),
     rulePgModel = new RulePgModel(),
     knex = await getKnexClient(),
     esClient = await Search.es(),
@@ -137,7 +141,7 @@ async function put(req, res) {
   const { params: { name }, body } = req;
 
   const apiRule = { ...body };
-  let newRule;
+  let translatedRule;
 
   if (name !== apiRule.name) {
     return res.boom.badRequest(`Expected rule name to be '${name}', but found`
@@ -145,40 +149,28 @@ async function put(req, res) {
   }
 
   try {
-    const oldRule = await ruleModel.get({ name });
+    const oldRule = await rulePgModel.get(knex, { name });
+    const oldApiRule = await translatePostgresRuleToApiRule(oldRule, knex);
 
     apiRule.updatedAt = Date.now();
-    apiRule.createdAt = oldRule.createdAt;
-
+    apiRule.createdAt = oldRule.created_at;
     // If rule type is onetime no change is allowed unless it is a rerun
     if (apiRule.action === 'rerun') {
-      return models.Rule.invoke(oldRule).then(() => res.send(oldRule));
+      return invokeRerun(oldApiRule).then(() => res.send(oldApiRule));
     }
+    const apiRuleWithTrigger = await updateRuleTrigger(oldApiRule, apiRule, knex);
+    const apiPgRule = await translateApiRuleToPostgresRuleRaw(apiRuleWithTrigger, knex);
 
-    const fieldsToDelete = Object.keys(oldRule).filter(
-      (key) => !(key in apiRule) && key !== 'createdAt'
-    );
-    const postgresRule = await translateApiRuleToPostgresRule(apiRule, knex);
-
-    try {
-      await createRejectableTransaction(knex, async (trx) => {
-        await rulePgModel.upsert(trx, postgresRule);
-        newRule = await ruleModel.update(oldRule, apiRule, fieldsToDelete);
-        await indexRule(esClient, newRule, process.env.ES_INDEX);
-      });
-    } catch (innerError) {
-      // Revert Dynamo record update if any write fails
-      await ruleModel.createRuleTrigger(oldRule);
-      await ruleModel.create(oldRule);
-      throw innerError;
-    }
-
-    return res.send(newRule);
+    await createRejectableTransaction(knex, async (trx) => {
+      const [pgRule] = await rulePgModel.upsert(trx, apiPgRule);
+      translatedRule = await translatePostgresRuleToApiRule(pgRule, knex);
+      await indexRule(esClient, translatedRule, process.env.ES_INDEX);
+    });
+    return res.send(translatedRule);
   } catch (error) {
     if (error instanceof RecordDoesNotExist) {
       return res.boom.notFound(`Rule '${name}' not found`);
     }
-
     throw error;
   }
 }
@@ -204,9 +196,11 @@ async function del(req, res) {
     process.env.ES_INDEX
   );
   let rule;
+  let apiRule;
 
   try {
     rule = await rulePgModel.get(knex, { name });
+    apiRule = await translatePostgresRuleToApiRule(rule, knex);
   } catch (error) {
     // If rule doesn't exist in PG or ES, return not found
     if (error instanceof RecordDoesNotExist) {
@@ -228,7 +222,7 @@ async function del(req, res) {
       index: process.env.ES_INDEX,
       ignore: [404],
     });
-    if (rule) await deleteRuleResources(knex, rule);
+    if (rule) await deleteRuleResources(knex, apiRule);
   });
 
   return res.send({ message: 'Record deleted' });